Implements parallel creation of shards
Makes use of the python3 specific concurrent.futures.
This works because the builder does not use the hash of shard N to produce the
shard N+1. Instead all the seeds are computed in advance.

I had to place a helper functions outside of the classes because multiprocessing
in python required picklable objects, which methods are not.