Fixes code to pass build tests
Had to move the cleanup check from generate_shards() to build() since shards are
returned from the former as a generator that get consumed in the later
Also fixes a bug where seeds used to generate the shards would be a successive
list of integer instead of proper hashed seeds.

I still get 2 errors related to registering addresses to the server, even with a
clean .stroj directory.